vm, ir: switch IR integer literals to dynamic width, move dynamic width integer type to IR crate
