+++ vm/src/value.rs
@@ -1,15 +1,4 @@
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Integer {
-    pub width: u8, pub signed: bool,
-    pub data: u64
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum Float {
-    F32(f32),
-    F64(f64)
-}
+pub use ir::{Integer, Float};
 
 /// Values have an implicit lifetime tied to the Heap they were allocated on
 #[derive(Clone, Debug, PartialEq)]
@@ -33,94 +22,3 @@ impl Value {
         }
     }
 }
-
-impl Integer {
-    pub fn new(width: u8, signed: bool, data: u64) -> Integer {
-        Integer { width, signed, data }
-    }
-
-    pub fn unsigned(width: u8, data: u64) -> Integer {
-        Integer { width, signed: false, data }
-    }
-    pub fn signed(width: u8, data: u64) -> Integer {
-        Integer { width, signed: true, data }
-    }
-
-    pub fn bitwise_negate(&self) -> Integer {
-        Integer {
-            data: !self.data,
-            ..*self
-        }
-    }
-
-    pub fn negate(&self) -> Integer {
-        assert!(self.signed);
-        Integer {
-            data: !self.data + 1,
-            ..*self
-        }
-    }
-}
-
-impl std::ops::Add for Integer {
-    type Output = Integer;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        assert!(self.width >= rhs.width);
-        assert!(self.signed == rhs.signed);
-        Integer {
-            data: self.data + rhs.data,
-            ..self
-        }
-    }
-}
-
-impl std::ops::Sub for Integer {
-    type Output = Integer;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        assert!(self.width >= rhs.width);
-        assert!(self.signed == rhs.signed);
-        Integer {
-            data: self.data - rhs.data,
-            ..self
-        }
-    }
-}
-
-impl std::ops::Mul for Integer {
-    type Output = Integer;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        assert!(self.width >= rhs.width);
-        assert!(self.signed == rhs.signed);
-        Integer {
-            data: self.data * rhs.data,
-            ..self
-        }
-    }
-}
-
-impl std::ops::Div for Integer {
-    type Output = Integer;
-
-    fn div(self, rhs: Self) -> Self::Output {
-        assert!(self.width >= rhs.width);
-        assert!(self.signed == rhs.signed);
-        Integer {
-            data: self.data / rhs.data,
-            ..self
-        }
-    }
-}
-
-impl Float {
-    pub fn width(&self) -> u8 {
-        match self {
-            Float::F32(_) => 32,
-            Float::F64(_) => 64,
-        }
-    }
-}
-
-
