+++ vm/test_modules/basic_math.s
@@ -6,14 +6,14 @@ Module(
     implementations: {},
     functions: {
         Symbol("start"): (
-            FunctionSignature(args: [], return_type: Int(width: 32, signed: true)),
+            FunctionSignature(args: [], return_type: Int(width: 64, signed: false)),
             FnBody(
                 max_registers: 2,
                 blocks: [
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Mul, Register(0), LiteralInt(3), LiteralInt(3)),
-                            BinaryOp(Sub, Register(1), Reg(Register(0)), LiteralInt(9)),
+                            BinaryOp(Mul, Register(0), LiteralInt(Integer(width: 64, signed: false, data: 3)), LiteralInt(Integer(width: 64, signed: false, data: 3))),
+                            BinaryOp(Sub, Register(1), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 9))),
                             Return(Reg(Register(1)))
                         ],
                         next_block: 0
