+++ vm/test_modules/branching.s
@@ -6,14 +6,14 @@ Module(
     implementations: {},
     functions: {
         Symbol("start"): (
-            FunctionSignature(args: [], return_type: Int(width: 32, signed: true)),
+            FunctionSignature(args: [], return_type: Int(width: 64, signed: false)),
             FnBody(
                 max_registers: 4,
                 blocks: [
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Mul, Register(0), LiteralInt(3), LiteralInt(3)),
-                            BinaryOp(Eq, Register(1), Reg(Register(0)), LiteralInt(9)),
+                            BinaryOp(Mul, Register(0), LiteralInt(Integer(width: 64, signed: false, data: 3)), LiteralInt(Integer(width: 64, signed: false, data: 3))),
+                            BinaryOp(Eq, Register(1), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 9))),
                             Br(cond: Reg(Register(1)), if_true: 1, if_false: 2),
                         ],
                         next_block: 999
@@ -21,8 +21,8 @@ Module(
 
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Mul, Register(2), LiteralInt(3), LiteralInt(3)),
-                            BinaryOp(NEq, Register(3), Reg(Register(0)), LiteralInt(9)),
+                            BinaryOp(Mul, Register(2), LiteralInt(Integer(width: 64, signed: false, data: 3)), LiteralInt(Integer(width: 64, signed: false, data: 3))),
+                            BinaryOp(NEq, Register(3), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 9))),
                             Br(cond: Reg(Register(3)), if_true: 2, if_false: 3)
                         ],
                         next_block: 999
@@ -30,14 +30,14 @@ Module(
 
                     BasicBlock(
                         instrs: [
-                            Return(LiteralInt(3333))
+                            Return(LiteralInt(Integer(width: 64, signed: false, data: 3333)))
                         ],
                         next_block: 999
                     ),
 
                     BasicBlock(
                         instrs: [
-                            Return(LiteralInt(0))
+                            Return(LiteralInt(Integer(width: 64, signed: false, data: 0)))
                         ],
                         next_block: 999
                     )
