+++ vm/test_modules/phi_nodes.s
@@ -6,14 +6,14 @@ Module(
     implementations: {},
     functions: {
         Symbol("start"): (
-            FunctionSignature(args: [], return_type: Int(width: 32, signed: true)),
+            FunctionSignature(args: [], return_type: Int(width: 64, signed: false)),
             FnBody(
                 max_registers: 5,
                 blocks: [
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Mul, Register(0), LiteralInt(3), LiteralInt(3)),
-                            BinaryOp(Eq, Register(1), Reg(Register(0)), LiteralInt(9)),
+                            BinaryOp(Mul, Register(0), LiteralInt(Integer(width: 64, signed: false, data: 3)), LiteralInt(Integer(width: 64, signed: false, data: 3))),
+                            BinaryOp(Eq, Register(1), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 9))),
                             Br(cond: Reg(Register(1)), if_true: 1, if_false: 2),
                         ],
                         next_block: 999
@@ -21,14 +21,14 @@ Module(
 
                     BasicBlock(
                         instrs: [
-                            LoadImm(Register(2), LiteralInt(0))
+                            LoadImm(Register(2), LiteralInt(Integer(width: 64, signed: false, data: 0)))
                         ],
                         next_block: 3
                     ),
 
                     BasicBlock(
                         instrs: [
-                            LoadImm(Register(3), LiteralInt(3333))
+                            LoadImm(Register(3), LiteralInt(Integer(width: 64, signed: false, data: 3333)))
                         ],
                         next_block: 3
                     ),
