+++ vm/test_modules/rec_call.s
@@ -8,9 +8,9 @@ Module(
         Symbol("fib"): (
             // (int, int, int) -> int
             FunctionSignature(args: [
-                    (Int(width: 32, signed: true), Symbol("n"))
+                    (Int(width: 64, signed: false), Symbol("n"))
                 ],
-                return_type: Int(width: 32, signed: true)
+                return_type: Int(width: 64, signed: false)
             ),
             // this code is sloppy and actually computes fib(n+2), oh well
             FnBody(
@@ -18,7 +18,7 @@ Module(
                 blocks: [
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Eq,  Register(1), Reg(Register(0)), LiteralInt(0)),
+                            BinaryOp(Eq,  Register(1), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 0))),
                             Br(cond: Reg(Register(1)), if_true: 1, if_false: 2)
                         ],
                         next_block: 999
@@ -26,15 +26,15 @@ Module(
 
                     BasicBlock(
                         instrs: [
-                            Return(LiteralInt(1))
+                            Return(LiteralInt(Integer(width: 64, signed: false, data: 1)))
                         ],
                         next_block: 999
                     ),
 
                     BasicBlock(
                         instrs: [
-                            BinaryOp(Sub, Register(2), Reg(Register(0)), LiteralInt(1)),
-                            BinaryOp(Sub, Register(3), Reg(Register(0)), LiteralInt(2)),
+                            BinaryOp(Sub, Register(2), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 1))),
+                            BinaryOp(Sub, Register(3), Reg(Register(0)), LiteralInt(Integer(width: 64, signed: false, data: 2))),
                             Call(Register(4), Path([Symbol("rec_call"), Symbol("fib")]), [ Reg(Register(2)) ]),
                             Call(Register(5), Path([Symbol("rec_call"), Symbol("fib")]), [ Reg(Register(3)) ]),
                             BinaryOp(Add, Register(6), Reg(Register(4)), Reg(Register(5))),
@@ -46,14 +46,14 @@ Module(
             )
         ),
         Symbol("start"): (
-            FunctionSignature(args: [], return_type: Int(width: 32, signed: true)),
+            FunctionSignature(args: [], return_type: Int(width: 64, signed: false)),
             FnBody(
                 max_registers: 2,
                 blocks: [
                     BasicBlock(
                         instrs: [
-                            Call(Register(0), Path([Symbol("rec_call"), Symbol("fib")]), [ LiteralInt(10) ]),
-                            BinaryOp(Sub, Register(1), LiteralInt(144), Reg(Register(0))),
+                            Call(Register(0), Path([Symbol("rec_call"), Symbol("fib")]), [ LiteralInt(Integer(width: 64, signed: false, data: 10)) ]),
+                            BinaryOp(Sub, Register(1), LiteralInt(Integer(width: 64, signed: false, data: 144)), Reg(Register(0))),
                             Return(Reg(Register(1)))
                         ],
                         next_block: 0
